cloud functions/cloud function 1/requirements.txt
python-telegram-bot
boto3